rename composite to base
