+++ packages/dbml-core/src/parse/mssql/fk_definition/index.js
@@ -1,6 +1,6 @@
 const P = require('parsimmon');
 const KP = require('../keyword_parsers');
-const { pDotDelimitedName, pIdentifier, pColumnNames } = require('../composite_parsers');
+const { pDotDelimitedName, pIdentifier, pColumnNames } = require('../base_parsers');
 const { makeList, makeNode } = require('../utils');
 const A = require('./actions');
